Added pub(crate) + cleanup
