+++ framework_crates/bones_framework/src/networking/online_lobby.rs
@@ -1,7 +1,7 @@
 #![allow(missing_docs)]
 
 use super::online::{
-    MatchmakerConnectionState, OnlineMatchmaker, OnlineMatchmakerRequest, OnlineMatchmakerResponse,
+    MatchmakerConnectionState,OnlineMatchmakerRequest, OnlineMatchmakerResponse,
     READ_TO_END_BYTE_COUNT,
 };
 use crate::{
@@ -10,11 +10,10 @@ use crate::{
     utils::BiChannelServer,
 };
 use bones_matchmaker_proto::{GameID, LobbyId, LobbyInfo, MatchmakerRequest, MatchmakerResponse};
-use iroh_net::NodeId;
 use std::sync::Arc;
 use tracing::info;
 
-pub async fn _resolve_list_lobbies(
+pub(crate) async fn resolve_list_lobbies(
     user_channel: &BiChannelServer<OnlineMatchmakerRequest, OnlineMatchmakerResponse>,
     matchmaker_connection_state: &mut MatchmakerConnectionState,
     game_id: GameID,
@@ -40,7 +39,7 @@ pub async fn _resolve_list_lobbies(
     Ok(())
 }
 
-pub async fn _resolve_create_lobby(
+pub(crate) async fn resolve_create_lobby(
     user_channel: &BiChannelServer<OnlineMatchmakerRequest, OnlineMatchmakerResponse>,
     matchmaker_connection_state: &mut MatchmakerConnectionState,
     lobby_info: LobbyInfo,
@@ -69,7 +68,7 @@ pub async fn _resolve_create_lobby(
     Ok(())
 }
 
-pub async fn _resolve_join_lobby(
+pub(crate) async fn resolve_join_lobby(
     user_channel: &BiChannelServer<OnlineMatchmakerRequest, OnlineMatchmakerResponse>,
     matchmaker_connection_state: &mut MatchmakerConnectionState,
     game_id: GameID,
@@ -141,46 +140,4 @@ pub async fn _resolve_join_lobby(
     }
 
     Ok(())
-}
-
-impl OnlineMatchmaker {
-    /// Sends a request to the matchmaking server to provide a list of all available lobbies for game_id. Response is read via `read_matchmaker_response()`.
-    pub fn list_lobbies(matchmaking_server: NodeId, game_id: GameID) -> anyhow::Result<()> {
-        super::online::ONLINE_MATCHMAKER
-            .try_send(OnlineMatchmakerRequest::ListLobbies {
-                id: matchmaking_server,
-                game_id,
-            })
-            .map_err(|e| anyhow::anyhow!("Failed to send list lobbies request: {}", e))?;
-        Ok(())
-    }
-
-    /// Sends a request to the matchmaking server to create a new lobby with the specified lobby_info.
-    pub fn create_lobby(matchmaking_server: NodeId, lobby_info: LobbyInfo) -> anyhow::Result<()> {
-        super::online::ONLINE_MATCHMAKER
-            .try_send(OnlineMatchmakerRequest::CreateLobby {
-                id: matchmaking_server,
-                lobby_info,
-            })
-            .map_err(|e| anyhow::anyhow!("Failed to send create lobby request: {}", e))?;
-        Ok(())
-    }
-
-    /// Sends a request to the matchmaking server to join a lobby with the specified game_id, lobby_id, and optional password.
-    pub fn join_lobby(
-        matchmaking_server: NodeId,
-        game_id: GameID,
-        lobby_id: LobbyId,
-        password: Option<String>,
-    ) -> anyhow::Result<()> {
-        super::online::ONLINE_MATCHMAKER
-            .try_send(OnlineMatchmakerRequest::JoinLobby {
-                id: matchmaking_server,
-                game_id,
-                lobby_id,
-                password,
-            })
-            .map_err(|e| anyhow::anyhow!("Failed to send join lobby request: {}", e))?;
-        Ok(())
-    }
 }
